requirements.txt
pyyaml
tqdm
rank-bm25
chromadb
sentence-transformers
openai
python-dotenv
